Even most reliably ActiveMediaSessionsListener
Exploring  the list of media session to search one that has transport controls, then grab that and set is as the media  session we can control with the lockscreen.
Previous behavior was:
Grab the first media session , and if this session has transport controls use it.

But what if the first media session didn't have Transport controls, but the second session did? (Assuming that the list currently has 2 or more media sessions) then the lockscreen will lose the ability to control playback of a session in particular.