Fixing a bug introduced in commit: fbcdb4e
Fixing a really strange bug where the Music Widget despite the music was playing it still got removed despite the code was being written correctly (Only despawn the music widget if the music was paused after 5 seconds), maybe a Xamarin bug, however it was fixed by using curly braces because it seems that using one-liner if's causes that if to not be evaluated, thus causing that bug.
Shrug xd.