Applying the BigPictureStyle (by using a new feature) BigpictureStyle is implemented as follows: Use the Image supplied to this notification and use it as Wallpaper as long as that notification is present (5 seconds) After that (new feature here) the Lockscreen will use the wallpaper it used before, for example(the default wallpaper)
The new feature is unstable so it's expected to not work too well, so far it works with Notifications with the BigPictureStyle style.

Changes:
LockscreenActivity: Now reacting to the event of 'the wallpaper has been cleared' a new event that simply indicates that whoever posted a Wallpaper is done with that(for example a Notification with BigPictureStyle) and any entity can post a new wallpaper; So the LockScreen will react to this as: If the previous Wallpaper was setted by myself (aka the Lockscreen) then resend the wallpaper used by the Lockscreen.

MusicFragment: doing the same as LockScreenActivity, Reacting to the event of 'the wallpaper has been cleared' and yadda yadda... If the previous Wallpaper was setted by myself (aka the Music Widget) then resend the current Album Art that is being held by this Music Widget to be used as Wallpaper

NotificationStyleApplier: Applying the BigPictureStyle.

CatcherHelper: Adding commented code thatll be uncommented(lol) this code is for making the Bundled/Stacked/Grouped Notifications to work, so far it does nothing, they are comments, xd.

CurrentWallpaperClearedEventArgs [Add]: Arguments for the event when the Wallpaper Is Cleared, arguments include a Property indicating which entity posted a wallpaper (it can be, LockScreen, MusicWidget or a Notification, for now)

WallpaperChangedEventArgs: Adding a new Property: WallpaperPoster, a flag that is used for identifying the current entity that's posting a Wallpaper.

WallpaperPosters: Enumeration with a list of possible entities that can post wallpapers to the lockscreen.

WallpaperPublisher: Implementing the feature that allows the BigPictureStyle to exist, it consists in the following:
When an entity post a new Wallpaper optionally it has the option to stablish a time in which that Wallpaper should be used in the lockscreen.
When this time passes the WallpaperPublisher clears this wallpaper and fires an event that Notifies all entities that the Wallpaper is cleared also along with a value indicating which entity was the previous one that posted a wallpaper before the wallpaper that was cleared was posted.
Now the entities should react in the following manner:
If this entity detects that the entity that posted the Wallpaper before the one that was deleted is itself, then this entity will do an action(such as resend the wallpaper that this entity wanted the lockscreen to use) in the case of the Lockscreen this is exactly what it does, allowing the Lockscreen to retrieve back the wallpaper it used (the LockScreen apart from acting as the place for posting Wallpapers, acts also as an entity that can post wallpapers to itself)