Implementing Better/Fixing the Notification Styles: Improved BigPictureStyle: now showing action buttons too. Adding the InboxStyle. Adding the BigTextStyle. Improving the MediaStyle: hiding the timeStamp as it seems that the timestamp is broken when a notification has this style. Adding the MessagingStyle.
NotificationSlave: the method overload for posting notification was ignoring the NotificationPriority parameter passed.

OpenNotification:
GetTextLines method fixed to be used by the InboxStyle style.
Add method GetBigText retrieves a really long text that is used in the BigTextStyle style.
Adding a test method for investigate about the Groups, Grouped notifications.

WARNING: These improvements also(for now) disables the capability of clearing the notification by pressing 'X' on it, I have broke it and I'm working on a fix for that.