Fixes #46
Checks if the the user has Unchecked the option of 'Music Widget Enabled' in the settings.
If is unchecked then stops the Floating Notification Service and loads the Notification Fragment again.
(This is a dirty fix, the ideal scenario would be to also check if the Floating Notification Serice is running in the first place, so I won't make unnecesary calls and try to stop something that's already stopped, same goes for loading the NotificationFragment, if it's already loaded, don't load it again.)