Notification Progress/Expandable Notification.
I started the implementation of the Expandable Notification, I don't know too much how to animate, so I just wrote some initial code that'll help me to get the ideas flowing, lol, and also I added a button on the notification widget that will do that action of toggling the Expanded/Collapsed state, for now it does nothing.

In the meantime, I now show on the lockscreen the progress of a notification that has to show a progress, for example a current download, this is fully implemented.

I added also a setting that lets me activate or deactivate debug stuff (like messages and toasts) directly on the Lockscreen settings, because sometimes I just want to use the lockscreen as normal, without annoying messages.

Added strings, and a new constant field in ConfigurationParameters, both are used for the 'debug' setting.

Deleting some unused code, like a couple of lines, I guess.