Solving a bug (not really, just a try catch)/Adding more test values.
When the floating notification widget is showing (a notification is being showed), Android will post a new notification of 'LiveDisplay is showing above other apps' (because of the current floating notification) click that Android notification on the list and it'll crash, the reason is when clicking that notification will cause the floating notification to be hidden for a brief moment. (in order to show the new clicked notification, in this case the Android one) causing the Android notification to be removed (because that notification only is showed if the Floating Notification is showed, and as I said, when clicking that floating Notification is hidden for a brief moment causing that Android notification to be removed) and now that is removed, the FloatingNotification will try to spawn and show that notification, but while doing that, it fails with 'OutOfbounds' Exception, because while loading data from the notification in that position, it gets removed. (I guess is a bug with the RecyclerView), the try catch will get the job done and the user won't notice it, because anyway that android notification gets removed so the user never gets to see it in the practice.

Adding more data to the test values in OpenNotification class, and also showing them in the Floating notification widget. (Previously only being showed on the Notification Fragment)