Solving an annoying bug! Previous behavior: If youre watching a notification while another notification was being updated constantly this notification that was being updated showed each time that got updated so causing you to press the notification you were viewing previously back again.
Now it works correctly, you watch the notification you want to watch.
(it still doesnt work in the floating notification, still working on it)

This change implied a huge refactor, but I did the regression tests, and everything works as expected along with this fix.

This  refactoring is all my fault and my lack of good design and planning lol, 1000 changes for a bug fix, the bright side is that now the code is better written, so it'll be easir to maintain in the future.