Improved Inline Response.
When youre writing a response the NotificationWidget will stay visible.
It won't stop lockscreen from blocking itself :/
you must write fast, for now.
(The callback of OnUserInteraction() does not work when the user is writing, so the Lockscreen will think that the user is not interacting with the lockscreen and will lock the screen, so I have to write a workaround for this, using wakelocks before the screen turns off, i think it won't affect battery duration as a User won't spend too much time writing a text, so those cases are edge ones)
When you send the inline response the keyboard will hide automatically (This has been just tested on my own device(Android 9), I hope it will work on other devices, :/)