Closes #44, Version Jump 0.9.0-alpha4!
The Fix was a hotfix, I tried a way to work with the notification priority/importance to be in harmony with Livedisplay blacklist but I coudn't

Fixing a behavior in Weather retrieval: it wasn't working, it worked but the first time.
Now the minimum interval for updating is 15 minutes (it was 10 but the Android API doesn't let me to do that)

Also Refactoring some code.

BootCompleteReceiver: update code to use my own apis to manage sharedpreferences.
IconFactory: refactor and warning removal.

(For all PreferencesFragments): change the parameter 'readAgain' in SetDefaultValues method from false to true to reread the defaultValues if not set already, because this causes some SharedPreferences to be not setted causing trouble in the app sometimes.

NotificationFragment: Reimplementing Awake here, in order to wake up the phone if a notification is important enough to do that

Blacklist: deleting non used code.
CatcherHelper: solving the issue #44

OpenNotification: Refactoring and correcting the method #MediaArtwork to work correctly in all supported API's

GrabWeatherJob: Stop using JobService in favor of WorkManager which is way easier to use and it works in Kitkat too!!

Weather: saving also the country of the city, because it was defaulting to US.