Version Jump 0.9.0-alpha5 Added more settings: Now you can hide the shortcuts (phone, camera) when the device is secured (pin, pattern, password, etc) the reason for this is that it makes no sense to have those shortcuts if you have to unlock the device anyway, shrug.
Changed behavior: Now when you send a test notification (from main settings) it also locks and turns off the screen, so the user can see when a new notification (test one) wakes up the device. (also fixing a bug in this thing of waking up, because the code was written but not tested)

Changing the settings to have more default values set:
Now the default for the following settings are:
Turn on new notification: yes.
Hide shrtcut when keyguard active: yes (added)
Turn off screen after last notification cleared: yes (added)
Fixing a bug where the 'Clear all' button in the lockscreen was shown even when all clearable notifications were gone, it was counting non removable notification as removable, so that's why it was showing up.