Version Jump 0.9.1-alpha1! Added a new feature: Now you can choose the music widget method. Android posses two methods to control media. 1. Use a token from a notification that represents a Media playing. 2. A list that holds all Sessions currently in the device.
Previous behavior was to use the best one from the list based on a simple if statement. (see ActiveMediaSessionsListener file/class)
Now you can also take it from the Notification.

This is a unstable feature so it can make the app crash.

Anyway the setting by default is 'Grab it from the Media sessions list'

I used this opportunity to refactor old code.