Version Jump 0.9.1-alpha3
Now the Floating notification has styles as the Notification Widget. (With this hange I got less lines of code! so it's so nice)

A Stable Wallpapers Settings section with no bugs.

To do: As the floating notification has style now it should be possible to answer directly from the notification as in the Notification Widget, however this isn't working, when you click the edit text, the keyboard doesn't show, so you can't write. :(