Tiny fixes. :)
If the Music Widget is present and the Album art setting says that the same should be blurred, then, if some other entity changed the wallpaper (Such as the weather) the Music Widget gained back control of the wallpaper, (as it should be) but the album art appeared back with no blur, (ignoring the user setting), that's fixed.

Coding style fixes.