Delivering Control Availability Checks to the Media Events Publishers, Media Fragment now displays controls depending on the controls available that are defined by the media events publishers. Deleting NotificationPosted event subscription, cuz I was using them to determine when to initialize the listening of the media events, but that's wrong, the media fragment should only listen, that's it, it's not its task to initialize anything. this allowed to detach this behavior, so now I have another Listener called the MediaControlsProviderService (including the MediaControlsProviderBroadcastReceiver to receive input from the notification), which shows the LiveDisplay Custom controls in a notification, to provide the full experience even if you're not in the lock screen activity, so you don't lose the usage flow, this service also reacts to the Media event publishers of course, in order to provide an uniform experience. now the IconFactory class is re-designed using the Builder pattern, so we can apply different things to the drawable we recover when using the constructor.
In Each MediaEvents publisher class we have a new event called ControlsAvailabilityChanged which informs listeners about the controls they can display.

in the Media Events publisher Lollipop I added a "Welcome" event functionality, which consists in, whenever a subscriber subscribes to any of the events then the subscriber receives an event immediately so it can keep up with the latest information available.

Media Events publishers initialization now is exclusively done by Catcher Service, which is the NotificationListenerService, this service is always started by android when the phone boots, so it's always guaranteed to be available to any subscribers, so the subscribers only have to worry about listening to them.

Media events publishers now have the capability to hold the notification that is associated with them, so we are left with more options in case we need them, and leave the listeners without that responsibility,

Added a new information to the Quick Glance Fragment which is the current messages count, which is a counter of how many Messaging Style notifications are present (Inbox Style if the Android version is Marshmallow or less), so the user won't have to dismiss the Lock Screen activity to see them, so we try to keep the user inside the app.

Adding related resources.