Adding a Quick Settings button to the LockScreenActivity, for now, two settings: Change the aspect ratio of the Current wallpaper Toggle on/off the ability to show the lock screen activity,
Subscribed to the QuickGlanceFragment messages button click, in order to display NotificationFragment when this event occurs.

Adding another condition to react when the UseWhenNoMediaPresent preference changes

QuickGlanceFragment: creating an event in the case the Messages button is clicked.

Added related resources.

UX: Upon Clicking the LockScreen Activity in a blank space causes the ripple animation to perform, that way we provide the user a hint that the lock screen should be pressed twice to exit, also pressing the back button to exit is re-enabled in case the user prefers this instead.