KeyguardPendingIntentMediator: Better exception handling, especially the CanceledException caused by a canceled PendingIntent, the SendPendingIntent method allows optionally to pass an alternate PendingIntent that might work better, this is used by the MediaEventsPublisherLollipop class to provide two possible pending intents it has, in order to ensure an activity will be started when the Open Related Activity Media command in this class is invoked by the MediaControlsBase class.
MediaEventsPublisherLollipop: Retrieving the OpenNotification associated with the media session this class holds, in order to obtain a PendingIntent that can be used to start the activity that owns the media session, in case the one we are provided by the session itself doesn't work.

Upon trying to open the related activity this class no longer checks for device lock status, as the KeyguardPendingIntentMediator already does that.